How do recursive types in Xtext work?
The problem is types are defined lexically in the grammar as a
left-recursive setup, particularly for optional types. For example,
Swift allows Int, Int? and Int?? as types. This can be defined in an
Xtext rule as:

Type:
  'Int' |
  Type '?' |
  'Bool';

However, the postfix operator ? (which can be doubled, tripled etc.)
results in a left-recursive grammar which Xtext doesn't permit directly.